tools/LogiAgent/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
autogen-agentchat==0.2.37
certifi==2024.8.30
charset-normalizer==3.4.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
exceptiongroup==1.2.2
filelock==3.16.1
FLAML==2.3.1
fsspec==2024.12.0
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.27.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
jiter==0.6.1
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
openai==1.53.0
packaging==24.1
pillow==11.1.0
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
pyparsing==3.2.0
pytest==8.3.4
python-dotenv==1.0.1
PyYAML==6.0.2
rank-bm25==0.2.2
regex==2024.9.11
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==3.3.1
sniffio==1.3.1
sympy==1.13.1
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
tqdm==4.66.6
transformers==4.48.0
typing_extensions==4.12.2
urllib3==2.2.3